+++ specification/ml/_types/Analysis.ts
@@ -19,6 +19,7 @@
 
 import { CategorizationAnalyzer } from '@ml/info/CategorizationAnalyzer'
 import { Field } from '@_types/common'
+import { long } from '@_types/Numeric'
 import { Time, TimeSpan } from '@_types/Time'
 import { Detector } from './Detector'
 
@@ -44,4 +45,16 @@ export class PerPartitionCategorization {
    * This setting can be set to true only if per-partition categorization is enabled. If true, both categorization and subsequent anomaly detection stops for partitions where the categorization status changes to warn. This setting makes it viable to have a job where it is expected that categorization works well for some partitions but not others; you do not pay the cost of bad categorization forever in the partitions where it works badly.
    */
   stop_on_warn?: boolean
+}
+
+export class AnalysisLimits {
+  categorization_examples_limit?: long
+  model_memory_limit: string
+}
+
+export class AnalysisMemoryLimit {
+  /**
+   * Limits can be applied for the resources required to hold the mathematical models in memory. These limits are approximate and can be set per job. They do not control the memory used by other processes, for example the Elasticsearch Java processes.
+   */
+  model_memory_limit: string
 }
