+++ specification/ml/_types/AnalysisMemoryLimit.ts
@@ -17,9 +17,4 @@
  * under the License.
  */
 
-export class AnalysisMemoryLimit {
-  /**
-   * Limits can be applied for the resources required to hold the mathematical models in memory. These limits are approximate and can be set per job. They do not control the memory used by other processes, for example the Elasticsearch Java processes.
-   */
-  model_memory_limit: string
-}
+
