+++ specification/ml/get_datafeed_stats/MlGetDatafeedStatsResponse.ts
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-import { DatafeedStats } from '@ml/_types/DatafeedStats'
+import { DatafeedStats } from '@ml/_types/Datafeed'
 import { long } from '@_types/Numeric'
 
 export class Response {
