+++ specification/ml/get_datafeeds/MlGetDatafeedsResponse.ts
@@ -21,5 +21,8 @@ import { Datafeed } from '@ml/_types/Datafeed'
 import { long } from '@_types/Numeric'
 
 export class Response {
-  body: { count: long; datafeeds: Datafeed[] }
+  body: {
+    count: long
+    datafeeds: Datafeed[]
+  }
 }
